requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
easyocr
opencv-python
pillow
numpy
pydantic
pyyaml
requests
google-generativeai
transformers
torch
pytest
python-dateutil
google-cloud-vision
groq